+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added the `/refinedstorage disk list <player>` command (raoulvdberge)
 - Added the `/refinedstorage network list <dimension>` command (raoulvdberge)
 - Added the `/refinedstorage network get <dimension> <pos>` command (raoulvdberge)
+- Added the `/refinedstorage network get <dimension> <pos> autocrafting list` command (raoulvdberge)
+- Added the `/refinedstorage network get <dimension> <pos> autocrafting get <id>` command (raoulvdberge)
 - Added JEI ghost ingredient dragging support (raoulvdberge)
 - Fixed text field not being focused in amount specifying screens (raoulvdberge)